Refactorizacion - Se modulariza la app. - Cambios en test (no terminados).
