+++ Sat.Recruitment.Api/Models/Result.cs
@@ -1,0 +1,8 @@
+﻿namespace Sat.Recruitment.Api.Models
+{
+    public class Result
+    {
+        public bool IsSuccess { get; set; }
+        public string Errors { get; set; }
+    }
+}
