+++ Sat.Recruitment.Api/Utilities/UserReader.cs
@@ -1,0 +1,17 @@
+﻿using System.IO;
+
+namespace Sat.Recruitment.Api.Utilities
+{
+    public class UserReader
+    {
+        public StreamReader ReadUsersFromFile()
+        {
+            var path = Directory.GetCurrentDirectory() + "/Files/Users.txt";
+
+            FileStream fileStream = new FileStream(path, FileMode.Open);
+
+            StreamReader reader = new StreamReader(fileStream);
+            return reader;
+        }
+    }
+}
