+++ Sat.Recruitment.Api/Validators/UserValidator.cs
@@ -1,0 +1,21 @@
+﻿namespace Sat.Recruitment.Api.Validators
+{
+    public class UserValidator
+    {
+        public void ValidateErrors(string name, string email, string address, string phone, ref string errors)
+        {
+            if (name == null)
+                //Validate if Name is null
+                errors = "The name is required";
+            if (email == null)
+                //Validate if Email is null
+                errors = errors + " The email is required";
+            if (address == null)
+                //Validate if Address is null
+                errors = errors + " The address is required";
+            if (phone == null)
+                //Validate if Phone is null
+                errors = errors + " The phone is required";
+        }
+    }
+}
